# crates/hs-admin/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: hs-admin API
  summary: The native administration and moderation API for the homeserver.
  version: 0.1.0
  description: 'The only API the management interface (track 16) talks to. See docs/rfcs/0004-admin-api.md for the full design rationale: resource modeling, naming, pagination, filtering, errors, idempotency, scopes, versioning and the event stream. This document is the contract; the hs-admin router''s contract tests assert the router and this document agree on paths and methods (RFC 0004 section 1, decision D15.16).'
  contact:
    name: 'Track 15: Admin API and modules'
    url: https://github.com/example/matrix-reimplement/blob/main/docs/workstreams/15-admin-api-and-modules.md
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  x-hs-contract: 1.0-draft
  x-hs-changelog:
  - version: 0.1.0
    date: '2026-09-18'
    notes: Initial draft covering every resource in RFC 0004 section 4. Not yet frozen; freezes at week 8 as contract 1.0.
servers:
- url: /api/v1
  description: Same origin as /_matrix on the client listener.
...
security:
- OAuth2: []
paths:
...
  /bridge-offerings/{type}/instances/{user_id}/files:
    post:
      summary: Render the files to run an instance elsewhere
      description: The instance's config, registration, a Compose service and a Kubernetes manifest (Secret and `Bridge`, for a cluster running the Myelin operator), with the instance's own tokens. Creates nothing.
      operationId: bridge_instances.files
      tags:
      - Bridges
      security:
      - OAuth2:
        - bridges:write
      parameters:
      - $ref: '#/components/parameters/BridgeTypeId'
      - $ref: '#/components/parameters/BridgeInstanceUser'
      responses:
        '200':
          description: The files.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BridgeInstanceFiles'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /rooms/{room_id}/events/{event_id}/context:
    get:
      summary: Get an event's context
      description: Reads message content, so it needs admin:read (a moderator's moderation:read does not reach it; see decision 0013), and every call is recorded in the audit log as `rooms.content.read`.
      operationId: rooms.events.context
      tags:
      - Rooms
      security:
      - OAuth2:
        - admin:read
      parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/EventId'
      - name: limit
        in: query
        schema:
          type: integer
          default: 10
      responses:
        '200':
          description: Context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventContext'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/InsufficientScope'
        '429':
          $ref: '#/components/responses/RateLimited'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: Tokens issued by 07's OAuth 2.0 issuer (authorization code with PKCE for operators, client credentials for automation). A legacy Matrix access token of a user with the server-administrator flag is also accepted and is treated as admin:read plus admin:write.
      flows:
        authorizationCode:
          authorizationUrl: https://example.org/_matrix/client/oauth2/authorize
          tokenUrl: https://example.org/_matrix/client/oauth2/token
          scopes:
            admin:read: Read every resource, including the audit log, configuration (secrets redacted), cluster and migration status. Satisfies every other :read scope.
            admin:write: Everything. Implies every other scope.
            bridges:read: Read appservices, bridge types, their health and backlog, and appservice tasks and events.
            bridges:write: bridges:read plus create, update, delete, pause, resume, replay, rotate tokens, export registrations, render bridge types.
            moderation:read: Read users, rooms, reports, media, their sub-resources, moderation tasks and events. Sessions' IP addresses and account data are redacted. Not message content (a room's timeline and its events), which needs admin:read (decision 0013).
            moderation:write: moderation:read plus suspend, lock, shadow-ban, deactivate, redact, reset password, logout, room block/delete/purge, quarantine and delete media, resolve reports, server notices.
        clientCredentials:
          tokenUrl: https://example.org/_matrix/client/oauth2/token
          scopes:
            admin:read: Read every resource, including the audit log, configuration (secrets redacted), cluster and migration status. Satisfies every other :read scope.
            admin:write: Everything. Implies every other scope.
            bridges:read: Read appservices, bridge types, their health and backlog, and appservice tasks and events.
            bridges:write: bridges:read plus create, update, delete, pause, resume, replay, rotate tokens, export registrations, render bridge types.
            moderation:read: Read users, rooms, reports, media, their sub-resources, moderation tasks and events. Sessions' IP addresses and account data are redacted. Not message content (a room's timeline and its events), which needs admin:read (decision 0013).
            moderation:write: moderation:read plus suspend, lock, shadow-ban, deactivate, redact, reset password, logout, room block/delete/purge, quarantine and delete media, resolve reports, server notices.
  parameters:
...
    RoomId:
      name: room_id
      in: path
      description: A Matrix room id, percent-encoded.
      required: true
      schema:
        type: string
      example: '%21abc%3Aexample.org'
    EventId:
      name: event_id
      in: path
      description: A Matrix event id, percent-encoded.
      required: true
      schema:
        type: string
...
    BridgeInstanceUser:
      name: user_id
      in: path
      description: The Matrix user the instance belongs to (`@alice:example.org`, URL-encoded), or `_` for a shared type's one instance.
      required: true
      schema:
        type: string
...
    BridgeTypeId:
      name: type
      in: path
      description: A bridge type catalog id.
      required: true
      schema:
        type: string
...
  responses:
...
    Unauthenticated:
      description: Missing, expired, revoked or malformed token (urn:hs:problem:unauthenticated).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer realm="hs-admin", error="invalid_token"
    InsufficientScope:
      description: The token lacks the operation's required scope (urn:hs:problem:insufficient-scope).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Bearer error="insufficient_scope", scope="..."
...
    NotFound:
      description: Unknown path or resource (urn:hs:problem:not-found).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Too many requests (urn:hs:problem:rate-limited).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds.
    Internal:
      description: Unexpected server error (urn:hs:problem:internal). Never includes internals.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unavailable:
      description: Owning replica unreachable, store unavailable, or read-only during cutover (urn:hs:problem:unavailable).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details, used for every /api/v1 error.
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
          description: urn:hs:problem:<slug> from the closed catalog in RFC 0004 section 3.5.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string
          description: ULID; matches X-Request-Id and tracing spans.
        errcode:
          type: string
          description: The Matrix error code, present whenever a Matrix client could observe the same failure.
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        retry_after_ms:
          type: integer
        required_scope:
          type: string
    ValidationError:
      type: object
      required:
      - pointer
      - detail
      properties:
        pointer:
          type: string
          description: JSON Pointer into the body, or param:<name> for a query parameter.
        detail:
          type: string
...
    StateEvent:
      type: object
      required:
      - event_id
      - type
      - state_key
      - sender
      - content
      - origin_server_ts
      properties:
        event_id:
          type: string
        type:
          type: string
        state_key:
          type: string
        sender:
          type: string
        content:
          type: object
          additionalProperties: true
        origin_server_ts:
          type: integer
    RoomEvent:
      type: object
      required:
      - event_id
      - room_id
      - type
      - sender
      - content
      - origin_server_ts
      - state_key
      - redacted
      properties:
        event_id:
          type: string
        room_id:
          type: string
        type:
          type: string
        sender:
          type: string
        content:
          type: object
          additionalProperties: true
        origin_server_ts:
          type: integer
        state_key:
          type:
          - string
          - 'null'
        redacted:
          type: boolean
          description: The event has been redacted; `content` is what redaction left.
        redacts:
          type:
          - string
          - 'null'
          description: For an `m.room.redaction`, the event it redacts.
    EventContext:
      type: object
      required:
      - event
      - events_before
      - events_after
      - state
      properties:
        event:
          $ref: '#/components/schemas/RoomEvent'
        events_before:
          type: array
          items:
            $ref: '#/components/schemas/RoomEvent'
        events_after:
          type: array
          items:
            $ref: '#/components/schemas/RoomEvent'
        state:
          type: array
          items:
            $ref: '#/components/schemas/StateEvent'
...
    BridgeInstanceFiles:
      type: object
      properties:
        config_yaml:
          type:
          - string
          - 'null'
        registration_yaml:
          type: string
        compose_yaml:
          type: string
        manifest_yaml:
          type: string
          description: A Secret with the files and a `Bridge` resource, for a cluster running the Myelin operator.
...